moved route validators to validation file, cleaned up code
