+++ backend/routes/api/session.js
@@ -2,47 +2,31 @@
 const express = require('express')
 const { Op } = require('sequelize');
 const bcrypt = require('bcryptjs');
-
 const { setTokenCookie, restoreUser } = require('../../utils/auth');
 const { User } = require('../../db/models');
-
-
-const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
-
+const { validateLogin } = require('../../utils/validation');
 
 const router = express.Router();
 
-const validateLogin = [
-  check('credential')
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage('Please provide a valid email or username.'),
-  check('password')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a password.'),
-  handleValidationErrors
-];
-
 //~ Restore session user
 router.get(
-    '/',
-    (req, res) => {
-      const { user } = req;
-      if (user) {
-        const safeUser = {
-          id: user.id,
-          firstName: user.firstName,
-          lastName: user.lastName,
-          email: user.email,
-          username: user.username,
-        };
-        return res.json({
-          user: safeUser
-        });
-      } else return res.json({ user: null });
-    }
-  );
+  '/',
+  (req, res) => {
+    const { user } = req;
+    if (user) {
+      const safeUser = {
+        id: user.id,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        email: user.email,
+        username: user.username,
+      };
+      return res.json({
+        user: safeUser
+      });
+    } else return res.json({ user: null });
+  }
+);
 
 //~ Log in
 router.post(
@@ -66,7 +50,7 @@ router.post(
       err.title = 'Login failed';
       err.errors = { credential: 'The provided credentials were invalid.' };
       return next(err);
-    }
+    };
 
     const safeUser = {
       id: user.id,
@@ -86,11 +70,11 @@ router.post(
 
 //~ Logout
 router.delete(
-'/',
-(_req, res) => {
-    res.clearCookie('token');
-    return res.json({ message: 'success' });
-}
+  '/',
+  (_req, res) => {
+      res.clearCookie('token');
+      return res.json({ message: 'success' });
+  }
 );
 
 module.exports = router;
