+++ backend/routes/api/spots.js
@@ -1,126 +1,16 @@
 //^ backend/routes/api/spots.js
 const express = require('express');
-const { Op, fn, col, where } = require('sequelize');
+const { Op, fn, col } = require('sequelize');
 const { Spot, Image, Review, User, Booking } = require('../../db/models');
 const { requireAuth, checkOwnership } = require('../../utils/auth');
 const formatDate = require('../api/utils/date-formatter');
-
-const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
+const { validateSpot, validateBooking, validateReview } = require('../../utils/validation');
 
 const router = express.Router();
-
-const validateSpot = [
-    check('address')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .withMessage('Street address is required'),
-    check('city')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .withMessage('City is required'),
-    check('state')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .withMessage('State is required'),
-    check('country')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .withMessage('Country is required'),
-    check('lat')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isNumeric()
-      .withMessage('Latitude must be numeric')
-      .custom((value, { req }) => {
-        const num = Number(value);
-        if (num < -90 || num > 90) {
-            throw new Error('Latitude must be within -90 and 90')
-        }
-        return true;
-      }),
-    check('lng')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isNumeric()
-      .withMessage('Latitude must be numeric')
-      .custom((value, { req }) => {
-        const num = Number(value);
-        if (num < -180 || num > 180) {
-            throw new Error('Latitude must be within -180 and 180')
-        }
-        return true;
-      }),
-    check('name')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isLength({ max: 50 })
-      .withMessage('Name must be less than 50 characters'),
-    check('description')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isLength({ min: 2, max: 250 })
-      .withMessage('Description is required'),
-    check('price')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isNumeric()
-      .withMessage('Price must be a number')
-      .custom((value, { req }) => {
-        const num = Number(value);
-        if (num < 0) {
-            throw new Error('Price per day must be a positive number')
-        }
-        return true;
-      }),
-    handleValidationErrors
-];
-
-const validateReview = [
-    check('review')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .withMessage('Review text is required')
-      .custom((value, { req }) => {
-        if (value.length > 250) {
-            throw new Error('Review must not exceed 250 charecters')
-        }
-        return true;
-      }),
-    check('stars')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isNumeric()
-      .withMessage('Stars must be a number')
-      .custom((value, { req }) => {
-        const num = Number(value);
-        if (num < 1 || num > 5) {
-            throw new Error('Stars must be an integer from 1 to 5')
-        }
-        return true;
-      }),
-    handleValidationErrors
-  ];
-
-  const validateBooking = [
-    check('startDate')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isDate()
-      .withMessage('startDate cannot be in the past'), //? custom validators?
-    check('endDate')
-      .exists({ checkFalsy: true })
-      .notEmpty()
-      .isDate()
-      .withMessage('endDate cannot be on or before startDate'), //? custom validators?
-    handleValidationErrors
-  ];
-
 
 //~ GET ALL SPOTS
 router.get('/', async (req, res) => {
     try {
-
         //^extract query params and assign defaults
         const {
             page = 1,
@@ -159,8 +49,7 @@ router.get('/', async (req, res) => {
             ...(maxLng && { lng: { [Op.lte]: maxLng } }),
             ...(minPrice && { price: { [Op.gte]: minPrice } }),
             ...(maxPrice && { price: { [Op.lte]: maxPrice } }),
-        }
-
+        };
 
         //^ get all spots including preview image data
         const spots = await Spot.findAll({
@@ -192,7 +81,6 @@ router.get('/', async (req, res) => {
 
         const responseData = spots.map(spot => {
             const spotObj = spot.get();
-
             delete spotObj.Images;
 
             return {
@@ -201,74 +89,68 @@ router.get('/', async (req, res) => {
                 updatedAt: formatDate(spot.updatedAt),
                 avgRating: avgRateMap[spot.id] || 1,
                 previewImage: spot.Images[0]?.url || null
-            }
+            };
         });
 
         return res.json({ Spots: responseData, page: parseInt(page, 10), size: parseInt(size, 10) });
-
     } catch (error) {
         console.error('Error fetching spots:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
     };
-
 });
 
 //~GET SPOTS OWNED BY CURRENT USER
 router.get('/current', requireAuth, async (req, res) => {
   try {
-      const spots = await Spot.findAll({
+    const spots = await Spot.findAll({
         where: {
-          ownerId: req.user.id
+            ownerId: req.user.id
         },
         include: [{
-          model: Image,
-          as: 'Images',
-          where: { preview: true },
-          attributes: ['url'],
-          limit: 1
+            model: Image,
+            as: 'Images',
+            where: { preview: true },
+            attributes: ['url'],
+            limit: 1
         }],
-      });
+    });
 
-      const reviewAverages = await Review.findAll({
-          attributes: ['spotId', [fn('SUM', col('stars')), 'sumStars'], [fn('COUNT', col('stars')), 'reviewCount']],
-          group: ['spotId']
-      });
+    const reviewAverages = await Review.findAll({
+        attributes: ['spotId', [fn('SUM', col('stars')), 'sumStars'], [fn('COUNT', col('stars')), 'reviewCount']],
+        group: ['spotId']
+    });
 
-      const avgRateMap = {};
-      reviewAverages.forEach( reviewAverage => {
-          const { spotId, sumStars, reviewCount } = reviewAverage.dataValues;
-          const avgRating = sumStars / reviewCount;
-          avgRateMap[spotId] = avgRating;
-      });
+    const avgRateMap = {};
+    reviewAverages.forEach( reviewAverage => {
+        const { spotId, sumStars, reviewCount } = reviewAverage.dataValues;
+        const avgRating = sumStars / reviewCount;
+        avgRateMap[spotId] = avgRating;
+    });
 
-      const responseData = spots.map(spot => {
-          const spotObj = spot.get();
+    const responseData = spots.map(spot => {
+        const spotObj = spot.get();
+        delete spotObj.Images;
 
-          delete spotObj.Images;
+        return {
+            ...spotObj,
+            createdAt: formatDate(spot.createdAt),
+            updatedAt: formatDate(spot.updatedAt),
+            avgRating: avgRateMap[spot.id] || 1,
+            previewImage: spot.Images[0]?.url || null
+        };
+    });
 
-          return {
-              ...spotObj,
-              createdAt: formatDate(spot.createdAt),
-              updatedAt: formatDate(spot.updatedAt),
-              avgRating: avgRateMap[spot.id] || 1,
-              previewImage: spot.Images[0]?.url || null
-          }
-      });
-
-      res.json({ Spots: responseData });
+    res.json({ Spots: responseData });
   } catch (error) {
     console.error('Error fetching spot:', error);
     return res.status(500).json({ error: 'Internal Server Error' });
   };
-
-
 });
-
 
 //~ GET SPOT DETAILS BY SPOT ID
 router.get('/:spotId', async (req, res) => {
-    const spotId = req.params.spotId;
     try {
+        const spotId = req.params.spotId;
         const spot = await Spot.findByPk(spotId, {
             include: [
                 {
@@ -287,7 +169,7 @@ router.get('/:spotId', async (req, res) => {
                 imageableType: 'spot',
                 imageableId: spotId
             }
-        })
+        });
 
         let sumStars = 1;
         let reviewCount = 0;
@@ -302,7 +184,7 @@ router.get('/:spotId', async (req, res) => {
                 const  { sumStars: totalStars, reviewCount: totalCount } = reviews[0].dataValues;
                 sumStars = totalStars;
                 reviewCount = totalCount;
-            }
+            };
 
         } catch (error) {
             console.error('Error fetching reviews:', error);
@@ -328,70 +210,73 @@ router.get('/:spotId', async (req, res) => {
                 id: owner.id,
                 firstName: owner.firstName,
                 lastName: owner.lastName
-            }
+            },
         };
 
         return res.json(responseData);
     } catch (error) {
         console.error('Error fetching spot:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-
+    };
 });
 
 //~ GET REVIEWS BY SPOTID
 router.get('/:spotId/reviews', async (req, res) => {
-
-    const reviews = await Review.findAll({
-        where: {
-            spotId: req.params.spotId
-        },
-        include: [
-            {
-                model: Image,
-                as: 'Images',
-                attributes: ['id', 'url'],
+    try {
+        const reviews = await Review.findAll({
+            where: {
+                spotId: req.params.spotId
             },
-            {
-                model: User,
-                as: 'User',
-                attributes: ['id', 'firstName', 'lastName'],
-            },
-        ],
-    });
+            include: [
+                {
+                    model: Image,
+                    as: 'Images',
+                    attributes: ['id', 'url'],
+                },
+                {
+                    model: User,
+                    as: 'User',
+                    attributes: ['id', 'firstName', 'lastName'],
+                },
+            ],
+        });
 
-    const responseData = reviews.map(review => {
-        const reviewObj = review.get();
+        const responseData = reviews.map(review => {
+            const reviewObj = review.get();
 
-        const reviewImages = reviewObj.Images.map( image => ({
-            id: image.id,
-            url: image.url,
-        }));
+            const reviewImages = reviewObj.Images.map( image => ({
+                id: image.id,
+                url: image.url,
+            }));
 
-        delete reviewObj.Images;
+            delete reviewObj.Images;
 
-        return {
-            ...reviewObj,
-            createdAt: formatDate(review.createdAt),
-            updatedAt: formatDate(review.updatedAt),
-            User: reviewObj.User ? {
-                id: reviewObj.User.id,
-                firstName: reviewObj.User.firstName,
-                lastName: reviewObj.User.lastName
-            } : null,
-            ReviewImages: reviewImages || null
-        }
-    });
+            return {
+                ...reviewObj,
+                createdAt: formatDate(review.createdAt),
+                updatedAt: formatDate(review.updatedAt),
+                User: reviewObj.User ? {
+                    id: reviewObj.User.id,
+                    firstName: reviewObj.User.firstName,
+                    lastName: reviewObj.User.lastName
+                } : null,
+                ReviewImages: reviewImages || null
+            };
+        });
 
-    res.json({ Reviews: responseData })
-})
+        res.json({ Reviews: responseData })
+    } catch (error) {
+        console.error('Error fetching review:', error);
+        return res.status(500).json({ error: 'Internal Server Error' });
+    };
+});
 
 //~GET BOOKINGS BY SPOT
 //! req auth && diff response depending on ownership of spot
 //! in route ownership check
 router.get('/:spotId/bookings', requireAuth, async (req, res) => {
     try {
-        const spot = await Spot.findByPk(req.params.spotId)
+        const spot = await Spot.findByPk(req.params.spotId);
         if (!spot) return res.status(404).json({ message: "Spot couldn't be found" });
 
         const bookings = await Booking.findAll({
@@ -402,7 +287,8 @@ router.get('/:spotId/bookings', requireAuth, async (req, res) => {
                     attributes: ['id', 'firstName', 'lastName']
                 }
             ]
-        })
+        });
+
         if (!bookings) return res.status(404).json({ message: `Bookings couldn't be found for this spot` });
 
         //^check ownership
@@ -417,9 +303,10 @@ router.get('/:spotId/bookings', requireAuth, async (req, res) => {
                     endDate: formatDate(booking.endDate).split(' ')[0],
                     createdAt: formatDate(spot.createdAt),
                     updatedAt: formatDate(spot.updatedAt)
-                }
-            })
-            res.json({ 'Bookings': ownedSpot })
+                };
+            });
+
+            res.json({ 'Bookings': ownedSpot });
         } else {
             const unOwnedSpot = bookings.map( booking => {
                 return {
@@ -431,12 +318,11 @@ router.get('/:spotId/bookings', requireAuth, async (req, res) => {
 
             res.json({ 'Bookings': unOwnedSpot });
         };
-
     } catch (error) {
         console.error('Error fetching bookings:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-})
+    };
+});
 
 //~ CREATE A SPOT
 //! req auth
@@ -444,6 +330,7 @@ router.post('/', requireAuth, validateSpot, async (req, res) => {
     try {
         const spotData = req.body;
         const owner = await User.findByPk(req.user.id);
+
         //^ enforce unique address, state, city combo
         const dupeCheck = await Spot.findOne({
             where: {
@@ -451,7 +338,7 @@ router.post('/', requireAuth, validateSpot, async (req, res) => {
                 city: spotData.city,
                 state: spotData.state
             }
-        })
+        });
 
         if (dupeCheck) return res.status(400).json({ message: 'Spot already exhists at this adress'});
 
@@ -466,20 +353,17 @@ router.post('/', requireAuth, validateSpot, async (req, res) => {
         spotObj.updatedAt = formatDate(spotObj.updatedAt);
 
         res.status(201).json(spotObj);
-
     } catch (error) {
         console.error('Error creating spot:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-
-})
+    };
+});
 
 //~EDIT A SPOT
 //! requires auth and ownership
 router.put('/:spotId', requireAuth, checkOwnership(Spot), validateSpot, async (req, res) => {
     try {
         const spot = await Spot.findByPk(req.params.spotId);
-
         spot.set({
             ...req.body
         });
@@ -487,35 +371,31 @@ router.put('/:spotId', requireAuth, checkOwnership(Spot), validateSpot, async (r
         spot.save({ validate: true });
 
         const spotObj = spot.get();
-
         spotObj.createdAt = formatDate(spotObj.createdAt);
         spotObj.updatedAt = formatDate(spotObj.updatedAt);
 
         res.status(200).json(spotObj);
-
     } catch (error) {
         console.error('Error updating spot:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-})
+    };
+});
 
 //~DELETE A SPOT
 //! requires auth and ownership
 router.delete('/:spotId', requireAuth, checkOwnership(Spot), async (req, res) => {
     try {
         const spot = await Spot.findByPk(req.params.spotId);
-
-        if (!spot) return res.status(404).json({ message: `Spot coudn't be found` })
+        if (!spot) return res.status(404).json({ message: `Spot coudn't be found` });
 
         spot.destroy();
 
         res.json({ message: 'Successfully deleted' });
-
     } catch (error) {
         console.error('Error deleteing spot:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-})
+    };
+});
 
 //~ ADD IMAGE TO SPOT BY SPOTID
 //! requires auth and ownership
@@ -523,7 +403,6 @@ router.post('/:spotId/images', requireAuth, checkOwnership(Spot, 'spotId'), asyn
     try {
         const spotData = req.body;
         const spotId = req.params.spotId;
-
         const spot = await Spot.findByPk(spotId);
         if(!spot) return res.status(404).json({ message: `Spot couldn't be found` });
 
@@ -534,22 +413,18 @@ router.post('/:spotId/images', requireAuth, checkOwnership(Spot, 'spotId'), asyn
 
         }, { validate: true });
 
-        const imageObj = newImage.get()
-
+        const imageObj = newImage.get();
         delete imageObj.imageableId;
         delete imageObj.imageableType;
         delete imageObj.createdAt;
         delete imageObj.updatedAt;
 
-        return res.status(201).json(imageObj)
-
-
+        return res.status(201).json(imageObj);
     } catch (error) {
         console.error('Error creating image:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-
-})
+    };
+});
 
 //~CREATE A REVIEW FOR A SPOT
 //! require auth
@@ -558,7 +433,6 @@ router.post('/:spotsId/reviews', requireAuth, validateReview, async (req, res) =
         const spot = await Spot.findByPk(req.params.spotsId);
         const user = await req.user.id;
         const review = req.body;
-
         if(!spot) return res.status(404).json({ message: `Spot couldn't be found` });
 
         //^ enforce unique user reviews
@@ -567,7 +441,7 @@ router.post('/:spotsId/reviews', requireAuth, validateReview, async (req, res) =
                 userId: req.user.id,
                 spotId: req.params.spotsId
             }
-        })
+        });
 
         if (dupeCheck) return res.status(500).json({ message: 'User already has a review for this spot'});
 
@@ -582,28 +456,26 @@ router.post('/:spotsId/reviews', requireAuth, validateReview, async (req, res) =
         newRevObj.updatedAt = formatDate(newRevObj.updatedAt);
 
         res.status(201).json(newRevObj);
-
-
     } catch (error) {
         console.error('Error creating review:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-})
+    };
+});
 
 //~CREATE A BOOKING FOR A SPOT
 //! req auth + spot may not belong to curr user
 router.post('/:spotId/bookings', requireAuth, validateBooking, async (req, res) => {
-    const spot = await Spot.findByPk(req.params.spotId);
-    if (!spot) return res.status(404).json({ message: `Spot couldn't be found`});
-
-    if (spot.ownerId === req.user.id) return res.status(403).json({ message: `Can't book your own spot`});
-
     try {
+        const spot = await Spot.findByPk(req.params.spotId);
+        if (!spot) return res.status(404).json({ message: `Spot couldn't be found`});
+        //^ check if user owns spot they are trying to book
+        if (spot.ownerId === req.user.id) return res.status(403).json({ message: `Can't book your own spot`});
+
         const newBooking = await spot.createBooking({
             spotId: spot.id,
             userId: req.user.id,
             ...req.body,
-        }, { validate: true })
+        }, { validate: true });
 
         const responseData = newBooking.get();
         responseData.startDate = formatDate(responseData.startDate).split(' ')[0];
@@ -615,9 +487,7 @@ router.post('/:spotId/bookings', requireAuth, validateBooking, async (req, res) 
     } catch (error) {
         console.error('Error creating booking:', error);
         return res.status(500).json({ error: 'Internal Server Error' });
-    }
-
+    };
 });
-
 
 module.exports = router;
