+++ backend/routes/api/users.js
@@ -1,67 +1,35 @@
 //^ backend/routes/api/users.js
 const express = require('express');
 const bcrypt = require('bcryptjs');
-
-const { setTokenCookie, requireAuth } = require('../../utils/auth');
+const { setTokenCookie } = require('../../utils/auth');
 const { User } = require('../../db/models');
-
-const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
+const { validateSignup } = require('../../utils/validation');
 
 const router = express.Router();
 
-const validateSignup = [
-  check('firstName')
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage('Please provide your first name'),
-  check('lastName')
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage('Please provide your last name'),
-  check('email')
-    .exists({ checkFalsy: true })
-    .isEmail()
-    .withMessage('Please provide a valid email.'),
-  check('username')
-    .exists({ checkFalsy: true })
-    .isLength({ min: 4 })
-    .withMessage('Please provide a username with at least 4 characters.'),
-  check('username')
-    .not()
-    .isEmail()
-    .withMessage('Username cannot be an email.'),
-  check('password')
-    .exists({ checkFalsy: true })
-    .isLength({ min: 6 })
-    .withMessage('Password must be 6 characters or more.'),
-  handleValidationErrors
-];
-
 //~ Sign up
 router.post(
-    '/',
-    validateSignup,
-    async (req, res) => {
-      const { firstName, lastName, email, password, username } = req.body;
-      const hashedPassword = bcrypt.hashSync(password);
-      const user = await User.create({ firstName, lastName, email, username, hashedPassword });
+  '/',
+  validateSignup,
+  async (req, res) => {
+    const { firstName, lastName, email, password, username } = req.body;
+    const hashedPassword = bcrypt.hashSync(password);
+    const user = await User.create({ firstName, lastName, email, username, hashedPassword });
 
-      const safeUser = {
-        id: user.id,
-        firstName: user.firstName,
-        lastName: user.lastName,
-        email: user.email,
-        username: user.username,
-      };
+    const safeUser = {
+      id: user.id,
+      firstName: user.firstName,
+      lastName: user.lastName,
+      email: user.email,
+      username: user.username,
+    };
 
-      await setTokenCookie(res, safeUser);
+    await setTokenCookie(res, safeUser);
 
-      return res.json({
-        user: safeUser
-      });
-    }
-  );
-
+    return res.json({
+      user: safeUser
+    });
+  }
+);
 
 module.exports = router;
